+++ playwright.config.mjs
@@ -31,6 +31,18 @@ export default defineConfig({
 			use: {
 				...devices["Desktop Chrome"]
 			}
+		},
+		{
+			name: "firefox",
+			use: {
+				...devices["Desktop Firefox"]
+			}
+		},
+		{
+			name: "webkit",
+			use: {
+				...devices["Desktop Safari"]
+			}
 		}
 	],
 	reporter: process.env["CI"]
@@ -49,5 +61,5 @@ export default defineConfig({
 		trace: "retain-on-failure",
 		video: "off"
 	},
-	workers: process.env["CI"] ? 2 : undefined
+	workers: 2
 });